* Remove size-limit on clauses

+++ src/pl-incl.h
@@ -1140,7 +1140,7 @@ struct clause
     unsigned long erased;		/* Generation I was erased */
   } generation;
 #endif /*O_LOGICAL_UPDATE*/
-  unsigned short	code_size;	/* size of ->codes */
+  unsigned int		code_size;	/* size of ->codes */
   unsigned short	variables;	/* # of variables for frame */
   unsigned short	prolog_vars;	/* # real Prolog variables */
 #ifdef O_SHIFT_STACKS
